create app header, update backgrounds
