+++ packages/client/src/app.tsx
@@ -3,7 +3,7 @@ import 'react-widgets/dist/css/react-widgets.css';
 import 'styles/app.scss';
 import classNames from 'classnames';
 import { ErrorBoundary } from 'react-error-boundary';
-import { useState, useEffect, ReactElement } from 'react';
+import { useEffect, ReactElement } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { useEthGasPrices } from 'hooks';
@@ -11,23 +11,12 @@ import { PendingTxProvider } from 'hooks/use-pending-tx';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import LandingContainer from 'containers/landing-container';
-import ConnectWalletModal from 'components/connect-wallet-modal';
-import { PageError, ModalError } from 'components/page-error';
+import { PageError } from 'components/page-error';
 
 import { WalletProvider } from 'hooks/use-wallet';
 
 function App(): ReactElement {
-    // ------------------ Initial Mount - API calls for first render ------------------
-
-    // const [allPairs, setAllPairs] = useState<AllPairsState>({
-    //     isLoading: true,
-    //     pairs: null,
-    //     lookups: null,
-    //     byLiquidity: null,
-    // });
-
     const gasPrices = useEthGasPrices();
-    const [showConnectWallet, setShowConnectWallet] = useState(false);
     // subscribe to the hook, will propogate to the nearest boundary
 
     const queryClient = new QueryClient();
@@ -35,34 +24,6 @@ function App(): ReactElement {
     useEffect(() => {
         document.body.classList.add('dark');
     }, []);
-    // useEffect(() => {
-    //     const fetchAllPairs = async () => {
-    //         // Fetch all pairs
-    //         const { data: pairsRaw, error } = await Uniswap.getTopPairs();
-
-    //         if (error) {
-    //             // we could not list pairs
-    //             console.warn(`Could not fetch top pairs: ${error}`);
-    //             debug.error = error;
-    //             setError(error);
-    //             return;
-    //         }
-
-    //         if (pairsRaw) {
-    //             const calculated = calculatePairRankings(pairsRaw);
-
-    //             setAllPairs({
-    //                 isLoading: false,
-    //                 pairs: calculated.pairs.map((p) => new UniswapPair(p)),
-    //                 lookups: calculated.pairLookups,
-    //                 byLiquidity: calculated.byLiquidity,
-    //             });
-    //         }
-    //     };
-
-    //     void fetchAllPairs();
-    //     // eslint-disable-next-line react-hooks/exhaustive-deps
-    // }, []);
 
     return (
         <ErrorBoundary
@@ -90,14 +51,6 @@ function App(): ReactElement {
                                 <div className='app-body' id='app-body'>
                                     <>
                                         <ErrorBoundary
-                                            FallbackComponent={ModalError}
-                                        >
-                                            <ConnectWalletModal
-                                                show={showConnectWallet}
-                                                setShow={setShowConnectWallet}
-                                            />
-                                        </ErrorBoundary>
-                                        <ErrorBoundary
                                             fallbackRender={({ error }) => (
                                                 <PageError errorMsg={error} />
                                             )}
@@ -106,17 +59,11 @@ function App(): ReactElement {
                                                 <Route path='/pools'>
                                                     <LandingContainer
                                                         gasPrices={gasPrices}
-                                                        setShowConnectWallet={
-                                                            setShowConnectWallet
-                                                        }
                                                     />
                                                 </Route>
                                                 <Route path='/'>
                                                     <LandingContainer
                                                         gasPrices={gasPrices}
-                                                        setShowConnectWallet={
-                                                            setShowConnectWallet
-                                                        }
                                                     />
                                                 </Route>
                                             </Switch>
