+++ packages/client/src/styles/colors.scss
@@ -2,7 +2,6 @@
     --purple100: #05001f;
     --purple90: #0b0044;
     --purple80: #18086b;
-    --purple75: yellow;
     --purple70: #2e1b91;
     --purple60: #4934b1;
     --purple50: #6652c9;
