+++ packages/client/src/styles/themes.scss
@@ -144,3 +144,10 @@
     --faceMoon: var(--green50);
     --faceDump: var(--red40);
 }
+/*
+background: rgb(253,182,54);
+background: -moz-linear-gradient(124deg, rgba(253,182,54,1) 0%, rgba(245,0,196,1) 100%);
+background: -webkit-linear-gradient(124deg, rgba(253,182,54,1) 0%, rgba(245,0,196,1) 100%);
+background: linear-gradient(124deg, rgba(253,182,54,1) 0%, rgba(245,0,196,1) 100%);
+filter: progid:DXImageTransform.Microsoft.gradient(startColorstr="#fdb636",endColorstr="#f500c4",GradientType=1);
+*/
